Move wchar/custom char overloads to xchar.h
