+++ test/os-test.cc
@@ -80,18 +80,6 @@ TEST(os_test, format_std_error_code) {
                         std::error_code(-42, fmt::system_category())));
 }
 
-TEST(os_test, format_std_error_code_wide) {
-  EXPECT_EQ(L"generic:42",
-            fmt::format(FMT_STRING(L"{0}"),
-                        std::error_code(42, std::generic_category())));
-  EXPECT_EQ(L"system:42",
-            fmt::format(FMT_STRING(L"{0}"),
-                        std::error_code(42, fmt::system_category())));
-  EXPECT_EQ(L"system:-42",
-            fmt::format(FMT_STRING(L"{0}"),
-                        std::error_code(-42, fmt::system_category())));
-}
-
 TEST(os_test, format_windows_error) {
   LPWSTR message = 0;
   auto result = FormatMessageW(
